backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
python-multipart
nibabel
opencv-python-headless
scikit-learn
scipy
python-dotenv
pydantic
httpx